+++ reactivestreams/subscription.py
@@ -2,6 +2,10 @@ from abc import ABCMeta, abstractmethod
 
 
 class Subscription(metaclass=ABCMeta):
+    """
+    Backpressure stream control.
+    """
+
     @abstractmethod
     def request(self, n: int):
         ...