+++ reactivestreams/publisher.py
@@ -4,6 +4,10 @@ from reactivestreams.subscriber import Subscriber
 
 
 class Publisher(metaclass=abc.ABCMeta):
+    """
+    Handles event for subscription to a subscriber
+    """
+
     @abc.abstractmethod
     def subscribe(self, subscriber: Subscriber):
         ...