+++ rsocket/disposable.py
@@ -1,0 +1,8 @@
+import abc
+
+
+class Disposable(metaclass=abc.ABCMeta):
+
+    @abc.abstractmethod
+    def dispose(self):
+        ...