+++ rsocket/streams/stream_from_async_generator.py
@@ -1,4 +1,6 @@
-from typing import AsyncGenerator, Tuple
+import asyncio
+from datetime import timedelta
+from typing import AsyncGenerator, Tuple, Callable
 
 from rsocket.async_helpers import async_range
 from rsocket.payload import Payload
@@ -7,8 +9,17 @@ from rsocket.streams.stream_from_generator import StreamFromGenerator
 
 
 class StreamFromAsyncGenerator(StreamFromGenerator):
+
+    def __init__(self,
+                 generator: Callable[[], AsyncGenerator[Tuple[Payload, bool], None]],
+                 delay_between_messages=timedelta(0),
+                 on_cancel=None,
+                 on_complete=None):
+        super().__init__(generator, delay_between_messages, on_cancel, on_complete)
+
     async def _start_generator(self):
-        self._iteration = self._generator().__aiter__()
+        self._generator: AsyncGenerator = self._generator_factory()
+        self._iteration = self._generator.__aiter__()
 
     async def _generate_next_n(self, n: int) -> AsyncGenerator[Tuple[Payload, bool], None]:
         is_complete_sent = False
@@ -21,3 +32,6 @@ class StreamFromAsyncGenerator(StreamFromGenerator):
                 if not is_complete_sent:
                     raise FinishedIterator()
                 return
+
+    def _cancel_generator(self):
+        asyncio.create_task(self._generator.aclose())