+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 ---------
 
+v0.4.7
+======
+- Send **cancel** to responders when requester disconnects and **error** to requesters when requester disconnects
+- Fix guide examples to properly cancel responders which use asyncio Task as value source
+- Rewrote guide statistics example to use generator instead of task
+
 v0.4.6
 ======
 - fire_and_forget now only removes the stream id when the future denoting the frame was sent, is done